+++ kubernetes/ingress.go
@@ -5,48 +5,97 @@
 package kubernetes
 
 import (
+	"errors"
 	"fmt"
+	"strings"
 
 	"github.com/tsuru/kubernetes-router/router"
 	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
+	typedV1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	typedV1Beta1 "k8s.io/client-go/kubernetes/typed/extensions/v1beta1"
+	v1 "k8s.io/client-go/pkg/api/v1"
 	"k8s.io/client-go/pkg/apis/extensions/v1beta1"
 )
 
-// IngressService manages ingresses in a Kubernetes cluster
+var (
+	// AnnotationsPrefix defines the common prefix used in the nginx ingress controller
+	AnnotationsPrefix = "nginx.ingress.kubernetes.io"
+	// AnnotationsNginx defines the common prefix used in the nginx ingress controller
+	AnnotationsNginx = map[string]string{"kubernetes.io/ingress.class": "nginx"}
+)
+
+// IngressService manages ingresses in a Kubernetes cluster that uses ingress-nginx
 type IngressService struct {
 	*BaseService
+	DefaultDomain string
 }
 
 // Create creates an Ingress resource pointing to a service
 // with the same name as the App
-func (k *IngressService) Create(appName string, _ router.Opts) error {
+func (k *IngressService) Create(appName string, routerOpts router.Opts) error {
+	var spec v1beta1.IngressSpec
+	var vhost string
 	client, err := k.ingressClient()
 	if err != nil {
 		return err
 	}
+	if len(routerOpts.Domain) > 0 {
+		vhost = routerOpts.Domain
+	} else {
+		vhost = fmt.Sprintf("%v.%v", appName, k.DefaultDomain)
+	}
+	spec = v1beta1.IngressSpec{
+		Rules: []v1beta1.IngressRule{
+			{
+				Host: vhost,
+				IngressRuleValue: v1beta1.IngressRuleValue{
+					HTTP: &v1beta1.HTTPIngressRuleValue{
+						Paths: []v1beta1.HTTPIngressPath{
+							{
+								Path: routerOpts.Route,
+								Backend: v1beta1.IngressBackend{
+									ServiceName: appName,
+									ServicePort: intstr.FromInt(defaultServicePort),
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+	}
+
 	i := v1beta1.Ingress{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:        ingressName(appName),
 			Namespace:   k.Namespace,
 			Labels:      map[string]string{appLabel: appName},
-			Annotations: make(map[string]string),
+			Annotations: k.Annotations,
 		},
-		Spec: v1beta1.IngressSpec{
-			Backend: &v1beta1.IngressBackend{
-				ServiceName: appName,
-				ServicePort: intstr.FromInt(defaultServicePort),
-			},
-		},
+		Spec: spec,
 	}
 	for k, v := range k.Labels {
 		i.ObjectMeta.Labels[k] = v
 	}
-	for k, v := range k.Annotations {
-		i.ObjectMeta.Annotations[k] = v
+	for k, v := range routerOpts.AdditionalOpts {
+		if !strings.Contains(k, "/") {
+			i.ObjectMeta.Annotations[annotationWithPrefix(k)] = v
+		} else {
+			i.ObjectMeta.Annotations[k] = v
+		}
 	}
+	if routerOpts.KubeLego {
+		i.Spec.TLS = []v1beta1.IngressTLS{
+			{
+				Hosts:      []string{i.Spec.Rules[0].Host},
+				SecretName: secretName(appName, i.Spec.Rules[0].Host),
+			},
+		}
+		i.ObjectMeta.Annotations["kubernetes.io/tls-acme"] = "true"
+	}
+
 	_, err = client.Create(&i)
 	if k8sErrors.IsAlreadyExists(err) {
 		return router.ErrIngressAlreadyExists
@@ -69,8 +118,8 @@ func (k *IngressService) Update(appName string, _ router.Opts) error {
 	if err != nil {
 		return err
 	}
-	ingress.Spec.Backend.ServiceName = service.Name
-	ingress.Spec.Backend.ServicePort = intstr.FromInt(int(service.Spec.Ports[0].Port))
+	ingress.Spec.Rules[0].IngressRuleValue.HTTP.Paths[0].Backend.ServiceName = service.Name
+	ingress.Spec.Rules[0].IngressRuleValue.HTTP.Paths[0].Backend.ServicePort = intstr.FromInt(int(service.Spec.Ports[0].Port))
 	_, err = ingressClient.Update(ingress)
 	return err
 }
@@ -136,12 +185,8 @@ func (k *IngressService) Get(appName string) (map[string]string, error) {
 	if err != nil {
 		return nil, err
 	}
-	var addr string
-	lbs := ingress.Status.LoadBalancer.Ingress
-	if len(lbs) != 0 {
-		addr = lbs[0].IP
-	}
-	return map[string]string{"address": addr}, nil
+
+	return map[string]string{"address": fmt.Sprintf("%v.%v", appName, ingress.Spec.Rules[0].Host)}, nil
 }
 
 func (k *IngressService) get(appName string) (*v1beta1.Ingress, error) {
@@ -164,12 +209,192 @@ func (k *IngressService) ingressClient() (typedV1Beta1.IngressInterface, error) 
 	return client.ExtensionsV1beta1().Ingresses(k.Namespace), nil
 }
 
+func (k *IngressService) secretClient() (typedV1.SecretInterface, error) {
+	client, err := k.getClient()
+	if err != nil {
+		return nil, err
+	}
+	return client.CoreV1().Secrets(k.Namespace), nil
+}
+
 func ingressName(appName string) string {
-	return appName + "-ingress"
+	return "kubernetes-router-" + appName + "-ingress"
+}
+
+func secretName(appName, certName string) string {
+	return "kubernetes-router-" + appName + "-" + certName + "-secret"
+}
+
+func annotationWithPrefix(suffix string) string {
+	return fmt.Sprintf("%v/%v", AnnotationsPrefix, suffix)
 }
 
 func (k *IngressService) swap(srcIngress, dstIngress *v1beta1.Ingress) {
 	srcIngress.Spec.Backend.ServiceName, dstIngress.Spec.Backend.ServiceName = dstIngress.Spec.Backend.ServiceName, srcIngress.Spec.Backend.ServiceName
 	srcIngress.Spec.Backend.ServicePort, dstIngress.Spec.Backend.ServicePort = dstIngress.Spec.Backend.ServicePort, srcIngress.Spec.Backend.ServicePort
 	k.BaseService.swap(&srcIngress.ObjectMeta, &dstIngress.ObjectMeta)
+}
+
+// AddCertificate adds certificates to app ingress
+func (k *IngressService) AddCertificate(appName string, certCname string, cert router.CertData) error {
+	ingressClient, err := k.ingressClient()
+	if err != nil {
+		return err
+	}
+	secret, err := k.secretClient()
+	if err != nil {
+		return err
+	}
+	ingress, err := k.get(appName)
+	if err != nil {
+		return err
+	}
+
+	tlsSecret := v1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        secretName(appName, certCname),
+			Namespace:   k.Namespace,
+			Labels:      map[string]string{appLabel: appName},
+			Annotations: make(map[string]string),
+		},
+		Type: "kubernetes.io/tls",
+		StringData: map[string]string{
+			"tls.key": cert.Key,
+			"tls.crt": cert.Certificate,
+		},
+	}
+	retSecret, err := secret.Create(&tlsSecret)
+	if err != nil {
+		return err
+	}
+
+	ingress.Spec.TLS = append(ingress.Spec.TLS,
+		[]v1beta1.IngressTLS{
+			{
+				Hosts:      []string{certCname},
+				SecretName: retSecret.Name,
+			},
+		}...)
+	_, err = ingressClient.Update(ingress)
+	return err
+}
+
+// GetCertificate get certificates from app ingress
+func (k *IngressService) GetCertificate(appName string, certCname string) (*router.CertData, error) {
+	secret, err := k.secretClient()
+	if err != nil {
+		return nil, err
+	}
+
+	retSecret, err := secret.Get(secretName(appName, certCname), metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+
+	certificate := fmt.Sprintf("%s", retSecret.Data["tls.crt"])
+	key := fmt.Sprintf("%s", retSecret.Data["tls.key"])
+	return &router.CertData{Certificate: certificate, Key: key}, err
+}
+
+// RemoveCertificate delete certificates from app ingress
+func (k *IngressService) RemoveCertificate(appName string, certName string) error {
+	ingressClient, err := k.ingressClient()
+	if err != nil {
+		return err
+	}
+	ingress, err := k.get(appName)
+	if err != nil {
+		return err
+	}
+	secret, err := k.secretClient()
+	if err != nil {
+		return err
+	}
+
+	ingress.Spec.TLS = nil
+	_, err = ingressClient.Update(ingress)
+	if err != nil {
+		return err
+	}
+
+	err = secret.Delete(secretName(appName, certName), &metav1.DeleteOptions{})
+
+	return err
+}
+
+// SetCname adds CNAME to app ingress
+func (k *IngressService) SetCname(appName string, cname string) error {
+	ingressClient, err := k.ingressClient()
+	if err != nil {
+		return err
+	}
+	ingress, err := k.get(appName)
+	if err != nil {
+		return err
+	}
+
+	annotations := ingress.GetAnnotations()
+	aliases, ok := annotations[annotationWithPrefix("server-alias")]
+	if !ok {
+		aliases = cname
+	} else {
+		aliasesArray := strings.Split(aliases, " ")
+		for _, v := range aliasesArray {
+			if strings.Compare(v, cname) == 0 {
+				return errors.New("cname already exists")
+			}
+		}
+		aliasesArray = append(aliasesArray, []string{cname}...)
+		aliases = strings.Join(aliasesArray, " ")
+	}
+	annotations[annotationWithPrefix("server-alias")] = aliases
+	ingress.SetAnnotations(annotations)
+
+	_, err = ingressClient.Update(ingress)
+
+	return err
+}
+
+// GetCnames get CNAMEs from app ingress
+func (k *IngressService) GetCnames(appName string) (*router.CnamesResp, error) {
+	ingress, err := k.get(appName)
+	if err != nil {
+		return nil, err
+	}
+
+	aliases, ok := ingress.GetAnnotations()[annotationWithPrefix("server-alias")]
+	if !ok {
+		return &router.CnamesResp{}, err
+	}
+
+	return &router.CnamesResp{Cnames: strings.Split(aliases, " ")}, err
+}
+
+// UnsetCname delete CNAME from app ingress
+func (k *IngressService) UnsetCname(appName string, cname string) error {
+	ingressClient, err := k.ingressClient()
+	if err != nil {
+		return err
+	}
+	ingress, err := k.get(appName)
+	if err != nil {
+		return err
+	}
+
+	annotations := ingress.GetAnnotations()
+	aliases := strings.Split(annotations[annotationWithPrefix("server-alias")], " ")
+
+	for index, value := range aliases {
+		if strings.Compare(value, cname) == 0 {
+			aliases = append(aliases[:index], aliases[index+1:]...)
+			break
+		}
+	}
+
+	annotations[annotationWithPrefix("server-alias")] = strings.Join(aliases, " ")
+	ingress.SetAnnotations(annotations)
+
+	_, err = ingressClient.Update(ingress)
+
+	return err
 }
